requirements.txt
langchain
cohere 
dataset 
pinecone-client 
langchain 
datasets 
openai
torch